+++ app/db/models/__init__.py
@@ -1,2 +1,3 @@
 from app.db.models import deprecated
+from app.db.models import document
 from app.db.session import Base
