+++ app/db/models/deprecated/document.py
@@ -1,6 +1,8 @@
 import sqlalchemy as sa
 from sqlalchemy import UniqueConstraint
 
+from app.db.models.document.physical_document import Language
+
 from ..auditable import Auditable
 from .source import Source
 from .geography import Geography
@@ -202,16 +204,6 @@ class DocumentResponse(Base):  # noqa: D101
         sa.ForeignKey(Document.id, ondelete="CASCADE"),
         nullable=False,
     )
-
-
-class Language(Base):  # noqa: D101
-    __tablename__ = "language"
-
-    id = sa.Column(sa.SmallInteger, primary_key=True)
-    language_code = sa.Column(sa.CHAR(length=3), nullable=False, unique=True)
-    part1_code = sa.Column(sa.CHAR(length=2))
-    part2_code = sa.Column(sa.CHAR(length=3))
-    name = sa.Column(sa.Text)
 
 
 class DocumentLanguage(Base):
