+++ app/db/models/document/__init__.py
@@ -1,0 +1,1 @@
+from .physical_document import PhysicalDocument
